move bottom drawer child components to bottomDrawer folder. add tbody to stage table. clean up bottom drawer child component css
