+++ src/components/bottomDrawer/beerGarden/beerGarden.js
@@ -1,8 +1,8 @@
 import React, {Component} from 'react';
 import './beerGarden.css';
-import { breweries } from "../../redux/constants";
-import { Tabs, Tab } from '../../../node_modules/@material-ui/core';
-import { TabContainer } from '../../helpers';
+import { breweries } from "../../../redux/constants";
+import { Tabs, Tab } from '../../../../node_modules/@material-ui/core';
+import { TabContainer } from '../../../helpers';
 
 class BeerGarden extends Component {
