+++ src/components/bottomDrawer/festivalInfo/festivalInfo.js
