+++ src/components/bottomDrawer/kidsDeck/kidsDeck.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
 import './kidsDeck.css';
-import { Tabs, Tab } from '../../../node_modules/@material-ui/core';
-import { TabContainer } from '../../helpers';
+import { Tabs, Tab } from '../../../../node_modules/@material-ui/core';
+import { TabContainer } from '../../../helpers';
 
 class KidsDeck extends Component {
