+++ src/components/bottomDrawer/lutefisk/lutefisk.js
