+++ src/components/bottomDrawer/smokedSalmon/smokedSalmon.js
